modificado:    CiA301CommPort.cpp 	modificado:    CiA301CommPort.h 	modificado:    Cia402device.cpp 	modificado:    Cia402device.h 	modificado:    ObjectDictionary.h

+++ Cia402device.h
@@ -76,7 +76,7 @@ public:
 
     long StartNode();
 
-    long SetVelocity(uint32_t target);
+    long SetVelocity(double target);
 private:
 
     //methods
